Split into more modular files and hierarchy
