+++ autobencher/util.py
@@ -1,0 +1,14 @@
+class Authorization(object):
+    def __init__(self, username, password):
+        self._username = username
+        self._password = password
+
+    @property
+    def username(self):
+        return self._username
+
+    @property
+    def password(self):
+        return self._password
+
+
